perf(linter): dispatch react_perf rules only on JSXAttribute nodes
The four react_perf rules got their `Rule` implementation from a blanket
`impl<R> Rule for R where R: ReactPerfRule` in `utils/react_perf.rs`.
Because the `run` body lived outside the rule files, `oxc_linter_codegen`
could not detect the node types these rules match, generating
`NODE_TYPES: None` — so all four rules ran on every AST node of every
JSX/TSX file.

Replace the blanket impl with per-rule `Rule` impls whose `run` starts
with the `let AstKind::JSXAttribute(..) = node.kind() else` pattern that
codegen detects, delegating to a shared `run_react_perf_rule` helper.
The rules are now dispatched only to `JSXAttribute` nodes.
