+++ crates/oxc_linter/src/rules/react_perf/jsx_no_jsx_as_prop.rs
@@ -5,7 +5,12 @@ use oxc_macros::declare_oxc_lint;
 use oxc_semantic::SymbolId;
 use oxc_span::{GetSpan, Span};
 
-use crate::utils::{NativeAllowList, ReactPerfConfig, ReactPerfRule};
+use crate::{
+    AstNode, LintContext,
+    context::ContextHost,
+    rule::{DefaultRuleConfig, Rule},
+    utils::{NativeAllowList, ReactPerfConfig, ReactPerfRule, run_react_perf_rule},
+};
 
 #[derive(Debug, Default, Clone, Deserialize)]
 pub struct JsxNoJsxAsProp(Box<ReactPerfConfig>);
@@ -52,6 +57,23 @@ declare_oxc_lint!(
     version = "0.2.3",
     short_description = "Prevent JSX elements that are local to the current method from being used as values of JSX props.",
 );
+
+impl Rule for JsxNoJsxAsProp {
+    fn from_configuration(value: serde_json::Value) -> Result<Self, serde_json::error::Error> {
+        serde_json::from_value::<DefaultRuleConfig<Self>>(value).map(DefaultRuleConfig::into_inner)
+    }
+
+    fn run<'a>(&self, node: &AstNode<'a>, ctx: &LintContext<'a>) {
+        let AstKind::JSXAttribute(attr) = node.kind() else {
+            return;
+        };
+        run_react_perf_rule(self, attr, node, ctx);
+    }
+
+    fn should_run(&self, ctx: &ContextHost) -> bool {
+        ctx.source_type().is_jsx()
+    }
+}
 
 impl ReactPerfRule for JsxNoJsxAsProp {
     const MESSAGE: &'static str = "JSX attribute values should not contain other JSX.";
